# docs/swagger.yaml
basePath: /
definitions:
...
  common.ApiErr:
    properties:
      Error:
        type: string
    type: object
...
  services.BlockPrice:
    properties:
      blockTime:
        type: integer
      createdAt:
        type: string
      delay:
        type: integer
      id:
        type: integer
      price:
        type: number
    type: object
  services.CoinPriceView:
    properties:
      bigPrice:
        type: string
      coin:
        description: 目标币价
        type: string
      node:
        type: string
      price:
        type: number
      sign:
        description: Sign_Hash值由 Timestamp+Coin+VsCoin+BigPrice计算
        items:
          type: integer
        type: array
      timestamp:
        type: integer
      vsCoin:
        description: vs币价
        type: string
    type: object
...
  services.PriceView:
    properties:
      bigPrice:
        type: string
      code:
        description: 合约代码
        type: string
      msg:
        description: debug msg
        type: string
      node:
        type: string
      nodeAddress:
        type: string
      priceUsd:
        type: number
      sign:
        description: Sign_Hash值由 Timestamp,Code,BigPrice
        items:
          type: integer
        type: array
      timestamp:
        type: integer
    type: object
...
  services.TokenTotalSupply:
    properties:
      createAt:
        type: string
      dailyAmount:
        type: string
      id:
        type: integer
      token:
        type: string
      totalSupply:
        type: string
    type: object
...
info:
  contact:
    email: wuxiaofei@rechaintech.com
    name: 伍晓飞
  description: oracle-api接口文档.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://test.com/index.html
  title: oracle-api
  version: "1.0"
paths:
  /pub/coin_price/{coin}/{vs_coin}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取币价换算，多节点签名版
      operationId: CoinPriceSignHandler
      parameters:
      - default: eth
        description: 目标币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: coin
        required: true
        type: string
      - default: usd
        description: vs币价
        enum:
        - btc
        - aed
        - ars
        - aud
        - bch
        - bdt
        - bhd
        - bits
        - bmd
        - bnb
        - brl
        - byn
        - cad
        - chf
        - clp
        - cny
        - czk
        - dkk
        - dot
        - eos
        - eth
        - eur
        - gbp
        - hkd
        - huf
        - idr
        - ils
        - inr
        - jpy
        - krw
        - kwd
        - link
        - lkr
        - ltc
        - mmk
        - mxn
        - myr
        - ngn
        - nok
        - nzd
        - php
        - pkr
        - pln
        - rub
        - sar
        - sats
        - sek
        - sgd
        - thb
        - try
        - twd
        - uah
        - usd
        - vef
        - vnd
        - xag
        - xau
        - xdr
        - xlm
        - xrp
        - yfi
        - zar
        in: path
        name: vs_coin
        required: true
        type: string
      - default: 1620383144
        description: unix 秒数
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: CoinPriceView
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.CoinPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取币价换算，多节点签名版
      tags:
      - default
  /pub/dex/ftx_chart_prices/{coin_type}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取杠杆btc代币不同时间区间的价格图表信息
      operationId: FtxChartPricesHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - btc3x
        - eth3x
        - vix3x
        - gold3x
        - gold3s
        - gold3x
        - eur20x
        - ndx10x
        - govt20x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 10
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: 1
        description: 数据间隔值,表示多少个15分钟, 如:1表示15分钟间隔 2表示30分钟间隔 3表示45分钟间隔 ,96表示1天间隔 ；
        enum:
        - 1
        - 2
        - 3
        - 4
        - 96
        in: path
        name: interval
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取杠杆btc代币不同时间区间的价格图表信息
      tags:
      - default
  /pub/dex/ftx_price/{coin_type}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取ftx token价格信息
      operationId: FtxPriceSignHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - usd
        - btc3x
        - eth3x
        - vix3x
        - govt20x
        - gold3x
        - gold3s
        - eur20x
        - ndx10x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取ftx token价格信息
      tags:
      - default
  /pub/dex/lp_price/{pair}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取lp价格信息
      operationId: PairLpPriceSignHandler
      parameters:
      - default: 0x21b8065d10f73ee2e260e5b47d3344d3ced7596e
        description: token地址
        in: path
        name: pair
        required: true
        type: string
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: pair price view list
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.PriceView'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取lp价格信息
      tags:
      - default
  /pub/dex/pair/token_chart_prices/{pair}/{token}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 从Pair获取token不同时间区间的价格图表信息
      operationId: PairTokenDayPricesHandler
      parameters:
      - default: 0x4612b8de9fb6281f6d5aa29635cf5700148d1b67
        description: pair地址
        in: path
        name: pair
        required: true
        type: string
      - default: 0x5df42c20d79fe40b51aba8fe5c8aa6531a3c453b
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 10
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: day
        description: '数据间隔 15minite hour day 1w(1周) 1m (1月) '
        enum:
        - 15minite
        - hour
        - day
        - 1w
        - 1m
        in: path
        name: interval
        required: true
        type: string
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 从Pair获取token不同时间区间的价格图表信息
      tags:
      - Pair
  /pub/dex/pair/token_price/{pair}/{token}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 从Pair获取token价格信息
      operationId: PairTokenPriceSignHandler
      parameters:
      - default: 0xdfb8824b094f56b9216a015ff77bdb056923aaf6
        description: pair地址
        in: path
        name: pair
        required: true
        type: string
      - default: 0x011864d37035439e078d64630777ec518138af05
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 从Pair获取token价格信息
      tags:
      - Pair
  /pub/dex/stock_chart_prices/{coin_type}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取股票不同时间区间的价格图表信息
      operationId: StockChartPricesHandler
      parameters:
      - default: AAPL
        description: 股票类型
        enum:
        - AAPL
        - TSLA
        in: path
        name: coin_type
        required: true
        type: string
      - default: 100
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: 4
        description: 数据间隔值,表示多少个15分钟, 如:1表示15分钟间隔 ４表示60分钟间隔 ,96表示1天间隔 ；
        enum:
        - 1
        - 4
        - 96
        in: path
        name: interval
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取股票不同时间区间的价格图表信息
      tags:
      - default
  /pub/dex/token/token_chart_supply/{token}/{amount}/{timestamp}:
    get:
      consumes:
      - application/json
      description: token totalSupply daily data
      operationId: TokenChartSupplyHandler
      parameters:
      - default: 0x011864d37035439e078d64630777ec518138af05
        description: token
        in: path
        name: token
        required: true
        type: string
      - default: 10
        description: 请求数量
        in: path
        name: amount
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: totalSupply
          schema:
            $ref: '#/definitions/services.TokenTotalSupply'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取某个token的totalSupply的变化量
      tags:
      - Token
  /pub/dex/token/token_total_supply/{token}/{timestamp}:
    get:
      consumes:
      - application/json
      description: token totalSupply
      operationId: GetTokenSupplyHandler
      parameters:
      - default: 0x011864d37035439e078d64630777ec518138af05
        description: token
        in: path
        name: token
        required: true
        type: string
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: totalSupply
          schema:
            $ref: '#/definitions/services.TokenTotalSupply'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取某个token的totalSupply的变化量
      tags:
      - Token
  /pub/dex/token_chain_price/{token}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取token链上价格信息，使用节点从eth或bsc合约事件监听到的价格变化数据；token信息要提前在节点配制才能被监听
      operationId: TokenChainPriceSignHandler
      parameters:
      - default: 0xc61624355667e4d5ca9cee25ad339c990a90eaea
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取token链上价格信息
      tags:
      - default
  /pub/dex/token_chart_prices/{token}/{count}/{interval}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取token不同时间区间的价格图表信息
      operationId: TokenDayPricesHandler
      parameters:
      - default: 0x66a0f676479cee1d7373f3dc2e2952778bff5bd6
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 10
        description: 获取多少个数据点
        in: path
        name: count
        required: true
        type: integer
      - default: day
        description: '数据间隔 15minite hour day 1w(1周) 1m (1月) '
        enum:
        - 15minite
        - hour
        - day
        - 1w
        - 1m
        in: path
        name: interval
        required: true
        type: string
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.BlockPrice'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取token不同时间区间的价格图表信息
      tags:
      - default
  /pub/dex/token_day_datas/{token}/{days}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取token相应天数的统计图表信息
      operationId: TokenDayDatasHandler
      parameters:
      - default: 0x66a0f676479cee1d7373f3dc2e2952778bff5bd6
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 14
        description: 获取最近多少天的数据
        in: path
        name: days
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: stock info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            items:
              $ref: '#/definitions/services.TokenDayData'
            type: array
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取token相应天数的统计图表信息
      tags:
      - default
  /pub/dex/token_price/{token}/{data_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取token价格信息
      operationId: TokenPriceSignHandler
      parameters:
      - default: 0x66a0f676479cee1d7373f3dc2e2952778bff5bd6
        description: token地址
        in: path
        name: token
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价
        enum:
        - 1
        - 2
        in: path
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      - default: 0
        description: 调试
        in: query
        name: debug
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: token price info
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLDataPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取token价格信息
      tags:
      - default
  /pub/dic_config:
    get:
      consumes:
      - application/json
      description: 字典配制信息
      operationId: ConfigHandler
      parameters:
      - description: 配制json
        in: query
        name: is_using
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: config json
          schema:
            $ref: '#/definitions/common.RawDicConfig'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 获取字典配制信息
      tags:
      - default
    post:
      consumes:
      - application/json
      description: 修改字典配制信息
      operationId: ConfigHandler
      parameters:
      - description: 配制json
        in: body
        name: dic
        required: true
        schema:
          $ref: '#/definitions/common.RawDicConfig'
      produces:
      - application/json
      responses:
        "200":
          description: config json
          schema:
            $ref: '#/definitions/common.RawDicConfig'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ResBody'
      summary: 修改字典配制信息
      tags:
      - default
...
  /pub/internal/dex/ftx_price/{coin_type}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 获取ftx coin最近一小时最高最低价格信息
      operationId: FtxPriceHandler
      parameters:
      - default: btc3x
        description: ftx类型
        enum:
        - mvi2x
        - usd
        - btc3x
        - eth3x
        - vix3x
        - gold3x
        - gold3s
        - eur20x
        - ndx10x
        - govt20x
        - mvi2s
        - btc3s
        - eth3s
        - vix3s
        - gold10s
        - eur20s
        - ndx10s
        - govt20s
        in: path
        name: coin_type
        required: true
        type: string
      - default: 1
        description: 最高最低价１最高　２最低价 3平均价 4最后价
        enum:
        - 1
        - 2
        - 3
        - 4
        in: query
        name: data_type
        required: true
        type: integer
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Price View
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.HLPriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: 获取ftx coin最近一小时最高最低价格信息,内部单节点
      tags:
      - default
  /pub/internal/dex/lp_price/{pair}/{timestamp}:
    get:
      consumes:
      - application/json
      description: 内部单节点获取lp价格信息
      operationId: PairLpPriceHandler
      parameters:
      - default: 0x21b8065d10f73ee2e260e5b47d3344d3ced7596e
        description: pair地址
        in: path
        name: pair
        required: true
        type: string
      - default: 1620383144
        description: 当前时间的unix秒数,该字段未使用，仅在云存储上用于标识
        in: path
        name: timestamp
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: price vew
          headers:
            sign:
              description: 签名信息
              type: string
          schema:
            $ref: '#/definitions/services.PriceView'
        "500":
          description: 失败时，有相应测试日志输出
          schema:
            $ref: '#/definitions/common.ApiErr'
      summary: '获取lp价格信息,内部单节点:'
      tags:
      - default
...
swagger: "2.0"